requirements.txt
PyYAML>=5.1.1
requests>=2.22.0
Flask>=1.0.3
Flask-SQLAlchemy>=2.4.0